Finish SPE Solver
This solver is untested, with a few caveats. Because solving
imperfect-information extensive form games requires a solution concept like
course-correlated equilibria (which might be ok, but would require CFR), or a
sequential equilibrium (which I believe requires exponential-time worst case to
solve), we just treat the toy climate game as a perfect information extensive
form game, making the stage game a commitment game. This should be fine as long
as the stage game remains a tragedy of the commons and not something like a
Chicken-like game, as the pure strategy Nash equilibria are the same as the pure
strategy Stackleberg equilibria in a ToC game.